use pillow coming from typeOfContract and remove duplicate views
