+++ Projects/ChangeTier/Sources/Service/DemoImplementation/ChangeTierClientDemo.swift
@@ -222,7 +222,8 @@ public class ChangeTierClientDemo: ChangeTierClient {
                 subTitle: "Endast en rörlig del om 25% av skadekostnaden.",
                 premium: .init(amount: "999", currency: "SEK")
             ),
-            canEditTier: true
+            canEditTier: true,
+            typeOfContract: .seApartmentBrf
         )
     }
