+++ Projects/MoveFlow/Sources/View/MovingFlowConfirm.swift
@@ -34,9 +34,8 @@ struct MovingFlowConfirm: View {
                         },
                         displayItems: $0.displayItems.map({ .init(title: $0.displayTitle, value: $0.displayValue) }),
                         insuranceLimits: $0.insurableLimits,
-                        onLimitTap: { limit in
-
-                        }
+                        onLimitTap: { limit in },
+                        typeOfContract: $0.contractType
                     )
                 })
