+++ Projects/hCoreUI/Sources/Views/ContractInformation.swift
@@ -1,12 +1,23 @@
 import SwiftUI
 
-struct ContractInformation: View {
+public struct ContractInformation: View {
     let displayName: String?
     let exposureName: String?
+    let pillowImage: UIImage?
 
-    var body: some View {
+    public init(
+        displayName: String?,
+        exposureName: String?,
+        pillowImage: UIImage?
+    ) {
+        self.displayName = displayName
+        self.exposureName = exposureName
+        self.pillowImage = pillowImage
+    }
+
+    public var body: some View {
         HStack(spacing: .padding12) {
-            Image(uiImage: hCoreUIAssets.pillowHome.image)
+            Image(uiImage: pillowImage ?? hCoreUIAssets.pillowHome.image)
                 .resizable()
                 .frame(width: 48, height: 48)
             VStack(alignment: .leading, spacing: 0) {
